Kurtosis E2E: factor out common test setup and starlark helper
- Add a shared runStarlarkScript helper (starlark.go).
- Split common Kurtosis test setup into kurtosis_e2e_helper.go.
- Bound enclave teardown with its own 1m context.
